Task2: Implemented find script by id routine

+++ src/main/java/com/brasiliwood/cooperfilmes/infrastructure/repository/MovieScriptRepository.java
@@ -3,5 +3,5 @@ package com.brasiliwood.cooperfilmes.infrastructure.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MovieScriptRepository
-        extends JpaRepository<MovieScriptEntity, Long> {
+        extends JpaRepository<MovieScriptEntity, Integer> {
 }
